change(ui): handle sessions on background, and other improvements

+++ tracker/src/main/java/com/openreplay/tracker/listeners/NetworkListener.kt
@@ -104,8 +104,8 @@ fun sendNetworkMessage(
         URL = url,
         request = requestJSON,
         response = responseJSON,
-        status = status.toULong(),
-        duration = duration.toULong()
+        status = status,
+        duration = duration
     )
 
     MessageCollector.sendMessage(message)
